requirements.txt
pynput
pytorch